Day5/Practice/etl/requirements.txt
flask>=2.0.1
gunicorn>=20.1.0
google-cloud-storage>=2.5.0
google-cloud-bigquery>=2.34.3
pandas>=1.4.2
python-dotenv>=0.19.2